requirements.txt
asgiref==3.8.1
async-timeout==5.0.0
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.12.14
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
Django==5.1.2
django-redis==5.4.0
-e git+https://github.com/jrief/django-websocket-redis@3b26de8666b92ac8e1c8880775168f8edf746d3b#egg=django_websocket_redis
djangorestframework==3.15.2
et-xmlfile==1.1.0
gevent==24.10.3
greenlet==3.1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
mysqlclient==2.2.4
numpy==2.1.3
opencv-python==4.10.0.84
openpyxl==3.1.5
pillow==11.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.2.1
redis==5.2.0
requests==2.32.3
service-identity==24.2.0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.2
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.3
websockets==14.0
zope.event==5.0
zope.interface==7.1.1
